feat: Add custom logo support to splash screen
Introduces the ability to display custom logos on the application's splash screen. This enhancement allows users to personalize the app's appearance by providing a base64 encoded image string for their logo. If no custom logo is provided, a default icon will be used.

Includes logic to render either a default icon from a predefined map or a custom image based on the `logoStr` prop. Also adds a failsafe to allow manual skipping of the splash screen by tapping.