feat(StudentProfile): Add 'Add Resource' button and dialog
Introduced a new "Add Resource" button to the Student Profile page. This button will trigger a dialog (dialog implementation not shown in diff) for users to add new learning materials or links for a student.

Additionally, removed the 'MATERIALS' tab from the student profile as it is no longer relevant.